Fix GeomFile binary format read
pyFormex GeomFiles (.pgf) saved in binary format could not be read back,
due to two bugs (one in GeomFile, one in arraytools.readArray.
Both have been fixed now.

Exporting in binary pgf is added as an option to the geometry menu,
as the format has now become functional.

+++ pyformex/geomfile.py
@@ -438,7 +438,7 @@ class GeometryFile(object):
 
         ndim = 3
         x = readArray(self.fil,Float,(ncoords,ndim),sep=sep)
-        e = readArray(self.fil,Float,(nelems,nplex),sep=sep)
+        e = readArray(self.fil,Int,(nelems,nplex),sep=sep)
         if props:
             p = readArray(self.fil,Int,(nelems,),sep=sep)
         else:
